requirements.txt
flask == 1.1.2
gunicorn == 20.0
virtualenv == 20.0.21
tensorflow == 2.4.1
reportlab == 3.5.59
mysql-connector == 2.2.9
requests == 2.25.1
numpy == 1.19.5